Fixed a design error in __eq__ method of Card class. Determining equality by value alone was causing issues with removing cards from the undealt list of a deck. Changed __eq__ to compare both suit and value, and added a new method eqVal to determine when two cards have the same value.
Completely overhauled pokergui.py to deal cards and display player hands and the community cards in separate windows. Added clearBoard, clearHand, addCards, dealBoard, and dealPlayers functions, as well as state, deck, boardImages, boardLabels, faceupBoard, numPlayers, handWindows, hands, handImages, and handLabels global variables to accomplish this.

Created .png graphics for the 52 playing cards of a standard deck, as well as for the backside of a card.