Fixed a bug with Hand greater and less functions returning true for equal value hands. Fixed an issue with Hand comparisons resulting from earlier change to card == operator (modified comparison functions to compare values rather than cards to avoid this).
Replaced global variables and functions in pokergui.py with Application class. The global variables used previously are now attributes of the class and the functions are methods. Modified functions to take in an extra "self" argument and changed references accordingly. Moved creation of buttons and initial images to __init__ method of Application.