In cards.py, fixed an issue in findFlushes HandHelper method that was preventing flushes from being found if there were only 5 cards to make the flush with.
In pokerchips.py, imported cards.py and added a hand attribute to the Player class. This attribute tracks the Player's hand as a Hand object. Modified resolvePots to take in an optional board argument. If this extra argument is provided, resolvePots will automatically find winners for all the pots using HandHelper.findWinner and the hands of the Players, instead of prompting the user for input on the terminal.

In pokergui.py, modified dealPlayers method to set the hand attribute of each Player in the table. Modified chipCounts StringVars to start with "Chips: ".