(Hopefully) fixed an issue in pokerchips and pokergui that was causing issues when blinds would take players all-in.
Modified proceed Application method to stop dealing out the rest of the cards immediately if betting stops due to players folding or going all-in. If a hand is ended by folding, any remaining board cards are no longer dealt and we proceed to the next hand. If players can no longer bet because they are all-in, board cards are dealt as the Proceed button is pressed. Added allPlayersFolded attribute to track when betting should stop due to folding and handle this case appropriately. Added some operations in proceed to reset players to an unfolded state when a new hand is dealt. Removed unused bettingOngoing attribute, added a postRound reset of the current bet and pot main window string variables, and modified disableBetting function to reset slider to bottom when a player is not betting.